requirements.txt
pyTelegramBotAPI
python-dotenv
bottle
spotipy